requirements.txt
certifi==2020.6.20
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
six==1.15.0
